# Optimize `TextServerAdvanced::_add_features` by using iteration instead of `.values()` and `.keys()`