I profiled the launch of an empty project and noticed this little contributor:
```
122.00 ms  100,0 %	8.00 ms	   TextServerAdvanced::_shape_run(TextServerAdvanced::ShapedTextDataAdvanced*, long long, long long, hb_script_t, hb_direction_t, TypedArray<RID>, long long, long long, long long, long long, RID
```

Its contribution can be reduced by 5% (a few ms) by using an iterator pattern, removing these contributors:
```
5.00 ms   4,1 %	0 s	    Dictionary::values() const
2.00 ms   1,6 %	1.00 ms	    Dictionary::keys() const
1.00 ms   0,8 %	1.00 ms	    Array::~Array()
```

## Addendum
It would be interesting to change `glyphs` to `LocalVector` (from `Vector`) and resize it fittingly, avoiding reallocations, because that contribution is also large:
```
17.00 ms  13,9 %	0 s	    Vector<Glyph>::push_back(Glyph)
```

However, that's more effort, so it can be done some other time.