@bruvzg 2025-03-21
> It would be interesting to change glyphs to LocalVector (from Vector) and resize it fittingly, avoiding reallocations, because that contribution is also large

Probably, but in most cases exact size of `glyphs` is not known in advance.