@YYF233333 2025-03-21
> It would be interesting to change `glyphs` to `LocalVector` (from `Vector`) and resize it fittingly, avoiding reallocations, because that contribution is also large

I tried this once in #101950 (without reserve because I don't know the exact size needed). It works but I can't really explain the reason 😂 as pointed out at https://github.com/godotengine/godot/pull/101950#pullrequestreview-2571659697, `Vector` also does power of 2 resize.